Refactor master profile to allow configuring pe_repos using hiera
Before this change, the puppet master profile had a hardcoded list of pe_repo platforms. This change refactors the profile to retrieve an array of platforms from hiera. It then transforms the array to the full class name using regex and includes the resulting array of classes.

This change also removes some duplicate hiera keys